pyproject.toml
[tool.poetry]
name = "udemy-slackbot"
version = "0.1.0"
description = ""
authors = ["kondotaichi <ubjdij210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
gradio = "3.31.0"
python-dotenv = "1.0.0"
langchain = "0.0.172"
openai = "0.27.6"
unstructured = "0.6.6"
tabulate = "0.9.0"
pdf2image = "1.16.3"
pytesseract = "0.3.10"
chromadb = "0.3.23"
tiktoken = "0.4.0"
slack-bolt = "1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
